Cardio

Cardio.adaptive_hrv_filter

Cardio.detectRPeaks

Cardio.detrend

Cardio.adaptive_hrv_filter — Method

adaptivehrvfilter(signal::Vector{<:Real}; removeoutliers::Bool = true, replacenonnormal::Bool = true, adaptivecontrollingcoef::Real = 0.05, proportionalitylimit::Real = 10/100, outlierminzfactor::Real = 3, maxexcesshrv::Real = 20, physiologicalvalues::Tuple{Real, Real} = (200, 2000))

An addaptive filter for HRV data Based on: Wessel, N., Voss, A., Malberg, H., Ziehmann, Ch., Voss, H. U., Schirdewan, A., Meyerfeldt, U.,Kurths, J.: Nonlinear analysis of complex phenomena in cardiological data, Herzschr. Elektrophys., 11(3), 2000, 159-173, doi:10.1007/s003990070035.

Filters out unphysiological beats in the RR series with the ability to replace them.

Args:

'signal::Vector': HRV in ms

Keyword Args

'remove_outliers::Bool = true': option if non physiological outliers shall be removed

'replace_nonnormal::Bool = true': if true, non normal HRV values are replaced

'adaptivecontrollingcoef::Real = 0.05': ???

'proportionality_limit::Real = 10/100': ???

'outlierminzfactor::Real = 3': ???

'maxexcesshrv::Real = 20': ???

'physiological_values::Tuple{Real, Real} = (200, 2000)': Definition of the physiological values, scheme: (min, max)

Return

Return value depends on keyword args

julia> adaptive_hrv_filter(signal)

Cardio.detectRPeaks — Method

detectRPeaks(ecg::Vector{<:Number}, samplerate::Number)

TODO check minDist Find R peaks in ECG signals as specified by Benitez et al. See http://dx.doi.org/10.1016/S0010-4825(01)00009-9 for more information

Args:

'ecg::Vector{<:Real}': ECG data

'samplerate::Real': Sampling rate [Hz]

'minPeakDist::Real': minimum distance between consecutive peaks [s]

Return:

'res::Vector{Int64}': Vector containing the position of the R peaks in ecg, divide by samplerate to get values in a time base

Examples

julia> a = detectRPeaks(ecg, fs)
Vector{Int64} with ....

Cardio.detrend — Method

detrend(signal::Vector{<:Real}; p::Int = 1, coefs::Union{Vector{<:Real}, Nothing, Real} = nothing, return_coefs::Bool = false)

Detrend a signal by removing polynomial trend of order p using build in least squares. Choose p = 0 to remove only mean or input coefficents from previous detrending to detrend by those.

Args:

signal::Vector: Data Vector containing te signal

p::Int: order of polynomial

coefs::Union{Vector{<:Real}, Nothing, Real}: Coefficents to do the same detrending on different signal

return_coefs::Bool: if true returns coeficients detrendet by

Return:

newSignal: The detrended sigal

Or if return_coefs = true:

(newSignal, coefs): The detrended sigal and the estimated coefficents from order 0 to p

Examples

julia> signal = sin.([1:100;]) + 0.03 .* [1:100;]
julia> detrend(signal)
Vector{Float}