BRS

Several methods to estimate baroreflex sensitivity (BRS) have been defined, many of which are contained in this package. All methods require the following inputs:

RR: The RR Interval series in ms

SBP: The systolic blood pressure in mmHg

Several methods can be tweaked by using keywords, though the most common and recommended parameters are set as default.

Missing docstring.

Missing docstring for getBRS. Check Documenter's build log for details.

Sequence Method

Cardio.BRS.sme — Function

sme(RR::Vector{<:Real}, SBP::Vector{<:Real}; thresholdRR::Float64 = 4.0, thresholdSBP::Float64 = 1.0, seqLen::Int = 3, delay::Int = 1, minCor::Float64 = 0.8)

Calculate the BaroReflex Sensitivity (BRS) for a serious of RR intervals and respectiv systolic bloodpressure using the Sequence Method (SME).

Args:

'RR::Vector': Data Vector containing the RR intervals

'SBP::Vector': Data Vector containing the systolic bloodpressure

Keywords:

'thresholdRR::Float64': Threshold for change in RR interval to qualify for a valid sequence (literature: 4ms)

'thresholdSBP::Float64': Threshold for change in SBP to qualify for a valid sequence (literature: 1mmHg)

'seqLen::Int': Minimum length of a valid sequence (literature: 3)

'delay::Int': Delay between RR and SBP (literature: 1)

'minCor::Float64': The minimal correlation between RR and SBP in a sequence to qualify as a valid sequence

'BRS::Float64': The BRS as estimated by the Sequence Method

Return:

Returns a SME struct for which a plotting recipe is provided, so that found sequences can be inspected by calling plot() when using Plots.jl

Examples

julia> sme(collect(1:10:100), collect(1:10:100))
1.0

Cardio.BRS.SME — Type

Struct that stores all information related to the sequence method. The main result is stored in 'sBRS'. It can be plotted for visual inspection.

result = BRS.sme(input.RR, input.SBP)
plot(result)

xBRS Method

Cardio.BRS.xbrs — Function

xbrs(RR::Vector{<:Real}, SBP::Vector{<:Real};...)

Calculate the xBRS index for the assesment of baroreflex sensitivity. Based on Westerhof, B. E. et al. (2004). Time-domain cross-correlation baroreflex sensitivity: performance on the EUROBAVAR data set. Journal of hypertension, 22(7), 1371-1380

Keyword Arguments

minCor: Minimal significant correlation, dafaults to 0.632 (p = 0.05, two-sided for 10 s window)

tExcerpt: Length of the sliding window in seconds, defaults to 10 s

delays: Which delays, shift of the RR window shall be considered, defaults to 0:5

Return

Returns an xBRS structure, that can be be plotted for visual inspection.

Cardio.BRS.xBRS — Type

Struct that stores all information regarting the xBRS etimation. The final result is stored in 'xBRSg'. It can be plotted for visual inspection.

result = BRS.xbrs(input.RR, input.SBP)
plot(result)

RMSSD Ratio

Cardio.BRS.rmssdr — Function

rmssdr(RR::Vector{<:Real}, SBP::Vector{<:Real})

Returns the RMSSD ratio := RMSSD(RR) / RMSSD(SBP)