Detection

ECG

Cardio.detectRPeaks — Function

detectRPeaks(ecg::Vector{<:Real}, samplerate::Real; minPeakDist::Real = 0.360)

TODO check minDist Find R peaks in ECG signals as specified by Benitez et al. See http://dx.doi.org/10.1016/S0010-4825(01)00009-9 for more information

Args:

'ecg::Vector{<:Real}': ECG data

'samplerate::Real': Sampling rate [Hz]

'minPeakDist::Real': minimum distance between consecutive peaks [s]

Return:

'res::Vector{Int64}': Vector containing the position of the R peaks in ecg, divide by samplerate to get values in a time base

Examples

julia> a = detectRPeaks(ecg, fs)
Vector{Int64} with ....

plot(ecg, lab = "")
peaks = detectRPeaks(ecg, 250) # signal was sampled at 250 Hz
scatter!(peaks, ecg[peaks])

Cardio.getECGBaseline — Function

getECGBaseline(ecg::Vector{<:Real}, samplerate::Real)

Get the baseline of an ECG signal for baseline correction. Source: Advances in Cardiac Signal Processing - Acharya, U.R. and Suri, J. and Spaan, J.A.E. and Krishnan, S.M. and Technologies, B. - ISBN: 9783540366751 page: 58f. adaption by Jan F. Kraemer

Args:

'ecg::Vector{<:Number}': ECG signal

'samplerate::Number': Sampling rate

Examples

julia> a = getECGBaseline(ecg, samplerate)
Vector{Float64} with ....

plot(ecg, lab = "")
baseline = getECGBaseline(ecg, 250) # signal was sampled at 250 Hz
plot!(baseline, lab = "baseline")

Blood Pressure