General Functionality

Cardio.BRS.detrend

Cardio.BRS.rmssdr

Cardio.BRS.sme

Cardio.BRS.theilSenEstimator

Cardio.adaptive_hrv_filter

Cardio.detectRPeaks

Cardio.detrend

Cardio.getECGBaseline

Cardio.medfilt1

Cardio.theilSenEstimator

Cardio.adaptive_hrv_filter — Method

adaptivehrvfilter(signal::Vector{<:Real}; removeoutliers::Bool = true, replacenonnormal::Bool = true, adaptivecontrollingcoef::Real = 0.05, proportionalitylimit::Real = 10/100, outlierminzfactor::Real = 3, maxexcesshrv::Real = 20, physiologicalvalues::Tuple{Real, Real} = (200, 2000))

An addaptive filter for HRV data Based on: Wessel, N., Voss, A., Malberg, H., Ziehmann, Ch., Voss, H. U., Schirdewan, A., Meyerfeldt, U.,Kurths, J.: Nonlinear analysis of complex phenomena in cardiological data, Herzschr. Elektrophys., 11(3), 2000, 159-173, doi:10.1007/s003990070035.

Filters out unphysiological beats in the RR series with the ability to replace them.

Args:

'signal::Vector': HRV in ms

Keyword Args

'remove_outliers::Bool = true': option if non physiological outliers shall be removed

'replace_nonnormal::Bool = true': if true, non normal HRV values are replaced

'adaptivecontrollingcoef::Real = 0.05': ???

'proportionality_limit::Real = 10/100': ???

'outlierminzfactor::Real = 3': ???

'maxexcesshrv::Real = 20': ???

'physiological_values::Tuple{Real, Real} = (200, 2000)': Definition of the physiological values, scheme: (min, max)

Return

Return value depends on keyword args

julia> adaptive_hrv_filter(signal)

Cardio.detectRPeaks — Method

detectRPeaks(ecg::Vector{<:Real}, samplerate::Real; minPeakDist::Real = 0.360)

TODO check minDist Find R peaks in ECG signals as specified by Benitez et al. See http://dx.doi.org/10.1016/S0010-4825(01)00009-9 for more information

Args:

'ecg::Vector{<:Real}': ECG data

'samplerate::Real': Sampling rate [Hz]

'minPeakDist::Real': minimum distance between consecutive peaks [s]

Return:

'res::Vector{Int64}': Vector containing the position of the R peaks in ecg, divide by samplerate to get values in a time base

Examples

julia> a = detectRPeaks(ecg, fs)
Vector{Int64} with ....

Cardio.detrend — Method

detrend(signal::Vector{<:Real}; p::Int = 1, coefs::Union{Vector{<:Real}, Nothing, Real} = nothing, return_coefs::Bool = false)

Detrend a signal by removing polynomial trend of order p using build in least squares. Choose p = 0 to remove only mean or input coefficents from previous detrending to detrend by those.

Args:

signal::Vector: Data Vector containing te signal

p::Int: order of polynomial

coefs::Union{Vector{<:Real}, Nothing, Real}: Coefficents to do the same detrending on different signal

return_coefs::Bool: if true returns coeficients detrendet by

Return:

newSignal: The detrended sigal

Or if return_coefs = true:

(newSignal, coefs): The detrended sigal and the estimated coefficents from order 0 to p

Examples

julia> signal = sin.([1:100;]) + 0.03 .* [1:100;]
julia> detrend(signal)
Vector{Float}

Cardio.getECGBaseline — Method

getECGBaseline(ecg::Vector{<:Real}, samplerate::Real)

Get the baseline of an ECG signal for baseline correction. Source: Advances in Cardiac Signal Processing - Acharya, U.R. and Suri, J. and Spaan, J.A.E. and Krishnan, S.M. and Technologies, B. - ISBN: 9783540366751 page: 58f. adaption by Jan F. Kraemer

Args:

'ecg::Vector{<:Number}': ECG signal

'samplerate::Number': Sampling rate

Examples

julia> a = getECGBaseline(ecg, samplerate)
Vector{Float64} with ....

Cardio.medfilt1 — Method

medfilt1(x::Array{<:Real}; n::Int = 3, padding::String = "zeropad", dim::Int = -1)

Apply a median filter to a signal vector or array x, similar to Matlabs medfilt1. Using Heap based calculation of the median to increase performance for larger windows n.

Args:

'x::Array{<:Real}': Array containing real values

Keywords:

'n::Int': Window length. The Median at point i is defined as median(x[i-n+1:i])

'padding::String': Specifies how to deal with Endpoints. The modes 'zeropad' and 'truncate' are available, with the first as default.

'dim::Int': Specifies the dimension to be filtered along. As default the first non singleton dimension is chosen.

Return:

'Array{Float64,N}': Always type Float64 with the same length as the input x

Examples

julia> medfilt1(collect(1:10))
10-element Array{Float64,1}:
 1.0
 2.0
 3.0
 4.0
 5.0
 6.0
 7.0
 8.0
 9.0
 9.0

Cardio.theilSenEstimator — Method

theilSenEstimator(x, y)

Calculate the Teil Sen Estimator (median of all slopes m = (yⱼ - yᵢ)/(xⱼ - xᵢ )). Stable up to ~27% outliers

Args:

'x::Vector': Data Vector containing x values

'y::Vector': Data Vector containing y values

Return:

'(m, b)::Tuple': m represents slope, b the intersect

Examples

julia> theilSenEstimator(1:10, 1:10)
(1.0,0.0)