Utility Functionality

Cardio.medfilt1 — Function

medfilt1(x::Array{<:Real}; n::Int = 3, padding::String = "zeropad", dim::Int = -1)

Apply a median filter to a signal vector or array x, similar to Matlabs medfilt1. Using Heap based calculation of the median to increase performance for larger windows n.

Args:

'x::Array{<:Real}': Array containing real values

Keywords:

'n::Int': Window length. The Median at point i is defined as median(x[i-n+1:i])

'padding::String': Specifies how to deal with Endpoints. The modes 'zeropad' and 'truncate' are available, with the first as default.

'dim::Int': Specifies the dimension to be filtered along. As default the first non singleton dimension is chosen.

Return:

'Array{Float64,N}': Always type Float64 with the same length as the input x

Examples

julia> medfilt1(collect(1:10))
10-element Array{Float64,1}:
 1.0
 2.0
 3.0
 4.0
 5.0
 6.0
 7.0
 8.0
 9.0
 9.0

Cardio.adaptive_hrv_filter — Function

adaptivehrvfilter(signal::Vector{<:Real}; removeoutliers::Bool = true, replacenonnormal::Bool = true, adaptivecontrollingcoef::Real = 0.05, proportionalitylimit::Real = 10/100, outlierminzfactor::Real = 3, maxexcesshrv::Real = 20, physiologicalvalues::Tuple{Real, Real} = (200, 2000))

An addaptive filter for HRV data Based on: Wessel, N., Voss, A., Malberg, H., Ziehmann, Ch., Voss, H. U., Schirdewan, A., Meyerfeldt, U.,Kurths, J.: Nonlinear analysis of complex phenomena in cardiological data, Herzschr. Elektrophys., 11(3), 2000, 159-173, doi:10.1007/s003990070035.

Filters out unphysiological beats in the RR series with the ability to replace them.

Args:

'signal::Vector': HRV in ms

Keyword Args

'remove_outliers::Bool = true': option if non physiological outliers shall be removed

'replace_nonnormal::Bool = true': if true, non normal HRV values are replaced

'adaptivecontrollingcoef::Real = 0.05': ???

'proportionality_limit::Real = 10/100': ???

'outlierminzfactor::Real = 3': ???

'maxexcesshrv::Real = 20': ???

'physiological_values::Tuple{Real, Real} = (200, 2000)': Definition of the physiological values, scheme: (min, max)

Return

Return value depends on keyword args

julia> adaptive_hrv_filter(signal)