BRS

Cardio.BRS.detrend

Cardio.BRS.rmssdr

Cardio.BRS.sme

Cardio.BRS.theilSenEstimator

Cardio.BRS.detrend — Method

detrend(signal::Vector{<:Real}; p::Int = 1, coefs::Union{Vector{<:Real}, Nothing, Real} = nothing, return_coefs::Bool = false)

Detrend a signal by removing polynomial trend of order p using build in least squares. Choose p = 0 to remove only mean or input coefficents from previous detrending to detrend by those.

Args:

signal::Vector: Data Vector containing te signal

p::Int: order of polynomial

coefs::Union{Vector{<:Real}, Nothing, Real}: Coefficents to do the same detrending on different signal

return_coefs::Bool: if true returns coeficients detrendet by

Return:

newSignal: The detrended sigal

Or if return_coefs = true:

(newSignal, coefs): The detrended sigal and the estimated coefficents from order 0 to p

Examples

julia> signal = sin.([1:100;]) + 0.03 .* [1:100;]
julia> detrend(signal)
Vector{Float}

Cardio.BRS.rmssdr — Method

rmssdr(RR::Vector{<:Real}, SBP::Vector{<:Real})

Returns the RMSSD ratio

Cardio.BRS.sme — Method

sme(RR::Vector{<:Real}, SBP::Vector{<:Real}; thresholdRR::Float64 = 4.0, thresholdSBP::Float64 = 1.0, seqLen::Int = 3, delay::Int = 1, minCor::Float64 = 0.8)

Calculate the BaroReflex Sensitivity (BRS) for a serious of RR intervals and respectiv systolic bloodpressure using the Sequence Method (SME).

Args:

'RR::Vector': Data Vector containing the RR intervals

'SBP::Vector': Data Vector containing the systolic bloodpressure

Keywords:

'thresholdRR::Float64': Threshold for change in RR interval to qualify for a valid sequence (literature: 4ms)

'thresholdSBP::Float64': Threshold for change in SBP to qualify for a valid sequence (literature: 1mmHg)

'seqLen::Int': Minimum length of a valid sequence (literature: 3)

'delay::Int': Delay between RR and SBP (literature: 1)

'minCor::Float64': The minimal correlation between RR and SBP in a sequence to qualify as a valid sequence

'BRS::Float64': The BRS as estimated by the Sequence Method

Return:

Returns a SME struct for which a plotting recipe is provided, so that found sequences can be inspected by calling plot() when using Plots.jl

Examples

julia> sme(collect(1:10:100), collect(1:10:100))
1.0

Cardio.BRS.theilSenEstimator — Method

theilSenEstimator(x, y)

Calculate the Teil Sen Estimator (median of all slopes m = (yⱼ - yᵢ)/(xⱼ - xᵢ )). Stable up to ~27% outliers

Args:

'x::Vector': Data Vector containing x values

'y::Vector': Data Vector containing y values

Return:

'(m, b)::Tuple': m represents slope, b the intersect

Examples

julia> theilSenEstimator(1:10, 1:10)
(1.0,0.0)